Added AppThemeAdapter to display the list of app theme colors and selection feedback. Fixed the visual appearance of the overdueTime in the EditTaskFragment, added a fetchSignInMethodsForUser function the the FirebaseAccess to check the linked social accounts to a user account,
added a progressbar and overlay for the avatar in the homepage, and the main page.

Added a settings details fragment to display the details of each setting option

and changed the SettingsFragment to one with SlidingPaneLayout.

Created a SharedPreferencesManager to handle sharedPreferences operations.

Added a signin with google option to the SignInFragment but haven't properly implemented it.